Code, Documentation, and Notebook Cleanup and Improvement
This code update cleans up the code and improves the demonstration notebook but maintains the same algorithm functionality used in the original HEROS paper. Includes updates to the README, Requirements, code organization, and the demonstration notebook. Within the demonstration notebook and code-base we've added functions to generate model prediction explanations for individual predictions, new visualizations, and examples of generating model feature importance estimates and generating an ROC curve with HEROS predictions.